Refactor code and add boundary detection

+++ src/view.c
@@ -4,18 +4,13 @@
 #define RED 255, 0, 0, 255
 
 void render(GameState* gameState, Bird bird) {
-    SDL_SetRenderDrawColor(gameState->renderer, BLACK);
-    SDL_RenderClear(gameState->renderer);
+  SDL_SetRenderDrawColor(gameState->renderer, BLACK);
+  SDL_RenderClear(gameState->renderer);
 
-    SDL_Rect bird_rect = {
-        bird.x, 
-        bird.y, 
-        bird.width, 
-        bird.height
-        };
+  SDL_Rect bird_rect = {bird.x, (int)bird.y, bird.width, bird.height};
 
-    SDL_SetRenderDrawColor(gameState->renderer, RED);
-    SDL_RenderFillRect(gameState->renderer, &bird_rect);
+  SDL_SetRenderDrawColor(gameState->renderer, RED);
+  SDL_RenderFillRect(gameState->renderer, &bird_rect);
 
-    SDL_RenderPresent(gameState->renderer);
+  SDL_RenderPresent(gameState->renderer);
 }
